Weekly calories and today meals displayed. photo size reduced

+++ app/src/main/java/com/sks225/snapeat/model/AltMeasure.kt
@@ -2,7 +2,7 @@ package com.sks225.snapeat.model
 
 data class AltMeasure(
     val measure: String,
-    val qty: Int,
+    val qty: Double,
     val seq: Int,
     val serving_weight: Double
 )
